A simulation of stock exchanges under construction
This is a simulation of a trading order book that records the online
behaviors of buyers and sellers; a naive matching engine has been built
to find pairs of trades that can be fulfilled. I will further improve
its functionality over time and the goal is to build a full-blown
trading engine.